wp root change and plugins update

+++ web/wp-content/plugins/wpforms/pro/includes/class-entry.php
@@ -3,11 +3,7 @@
 /**
  * Entry DB class.
  *
- * @package    WPForms
- * @author     WPForms
- * @since      1.0.0
- * @license    GPL-2.0+
- * @copyright  Copyright (c) 2016, WPForms LLC
+ * @since 1.0.0
  */
 class WPForms_Entry_Handler extends WPForms_DB {
 
@@ -29,11 +25,13 @@ class WPForms_Entry_Handler extends WPForms_DB {
 	 * Get table columns.
 	 *
 	 * @since 1.0.0
+	 * @since 1.5.7 Added an `Entry Notes` column.
 	 */
 	public function get_columns() {
 
 		return array(
 			'entry_id'      => '%d',
+			'notes_count'   => '%d',
 			'form_id'       => '%d',
 			'post_id'       => '%d',
 			'user_id'       => '%d',
@@ -75,21 +73,83 @@ class WPForms_Entry_Handler extends WPForms_DB {
 	}
 
 	/**
-	 * Deletes an entry from the database, also removes entry meta.
+	 * Retrieve an entry from the database based on a given entry ID.
+	 *
+	 * @since 1.5.8
+	 *
+	 * @param int   $id   Entry ID.
+	 * @param array $args Additional arguments.
+	 *
+	 * @return object|null
+	 */
+	public function get( $id, $args = array() ) {
+
+		if ( ! isset( $args['cap'] ) && wpforms()->get( 'access' )->init_allowed() ) {
+			$args['cap'] = 'view_entry_single';
+		}
+
+		if ( ! empty( $args['cap'] ) && ! wpforms_current_user_can( $args['cap'], $id ) ) {
+			return null;
+		}
+
+		return parent::get( $id );
+	}
+
+	/**
+	 * Update an existing entry in the database.
+	 *
+	 * @since 1.5.8
+	 *
+	 * @param string $id    Entry ID.
+	 * @param array  $data  Array of columns and associated data to update.
+	 * @param string $where Column to match against in the WHERE clause. If empty, $primary_key
+	 *                      will be used.
+	 * @param string $type  Data type context.
+	 * @param array  $args  Additional arguments.
+	 *
+	 * @return bool|null
+	 */
+	public function update( $id, $data = array(), $where = '', $type = '', $args = array() ) {
+
+		if ( ! isset( $args['cap'] ) ) {
+			$args['cap'] = ( array_key_exists( 'viewed', $data ) || array_key_exists( 'starred', $data ) ) ? 'view_entry_single' : 'edit_entry_single';
+		}
+
+		if ( ! empty( $args['cap'] ) && ! wpforms_current_user_can( $args['cap'], $id ) ) {
+			return null;
+		}
+
+		return parent::update( $id, $data, $where, $type );
+	}
+
+	/**
+	 * Delete an entry from the database, also removes entry meta.
 	 *
 	 * Please note: successfully deleting a record flushes the cache.
 	 *
 	 * @since 1.1.6
 	 *
-	 * @param int $row_id Entry ID.
+	 * @param int   $row_id Entry ID.
+	 * @param array $args   Additional arguments.
 	 *
 	 * @return bool False if the record could not be deleted, true otherwise.
 	 */
-	public function delete( $row_id = 0 ) {
+	public function delete( $row_id = 0, $args = array() ) {
+
+		if ( ! isset( $args['cap'] ) ) {
+			$args['cap'] = 'delete_entry_single';
+		}
+
+		if ( ! empty( $args['cap'] ) && ! wpforms_current_user_can( $args['cap'], $row_id ) ) {
+			return false;
+		}
 
 		$entry  = parent::delete( $row_id );
 		$meta   = wpforms()->entry_meta->delete_by( 'entry_id', $row_id );
 		$fields = wpforms()->entry_fields->delete_by( 'entry_id', $row_id );
+
+		WPForms\Pro\Admin\DashboardWidget::clear_widget_cache();
+		WPForms\Pro\Admin\Entries\DefaultScreen::clear_widget_cache();
 
 		return ( $entry && $meta && $fields );
 	}
@@ -273,6 +333,7 @@ class WPForms_Entry_Handler extends WPForms_DB {
 	 * Get entries from the database.
 	 *
 	 * @since 1.0.0
+	 * @since 1.5.7 Added a `notes_count` argument to request the count of notes for each entry.
 	 *
 	 * @param array $args  Redefine query parameters by providing own arguments.
 	 * @param bool  $count Whether to just count entries or get the list of them. True to just count.
@@ -300,6 +361,7 @@ class WPForms_Entry_Handler extends WPForms_DB {
 			'date'          => '',
 			'date_modified' => '',
 			'ip_address'    => '',
+			'notes_count'   => false,
 			'orderby'       => 'entry_id',
 			'order'         => 'DESC',
 		);
@@ -308,6 +370,8 @@ class WPForms_Entry_Handler extends WPForms_DB {
 			'wpforms_entry_handler_get_entries_args',
 			wp_parse_args( $args, $defaults )
 		);
+
+		$fields_table = wpforms()->entry_fields->table_name;
 
 		/*
 		 * Modify the SELECT.
@@ -318,7 +382,7 @@ class WPForms_Entry_Handler extends WPForms_DB {
 			'wpforms_entry_handler_get_entries_select',
 			array(
 				'all'       => '*',
-				'entry_ids' => '`entry_id`',
+				'entry_ids' => "{$this->table_name}.entry_id",
 			)
 		);
 		if ( array_key_exists( $args['select'], $possible_select_values ) ) {
@@ -337,8 +401,7 @@ class WPForms_Entry_Handler extends WPForms_DB {
 		);
 
 		// Allowed int arg items.
-		$keys = array( 'entry_id', 'form_id', 'post_id', 'user_id', 'viewed', 'starred' );
-		foreach ( $keys as $key ) {
+		foreach ( array( 'entry_id', 'form_id', 'post_id', 'user_id', 'viewed', 'starred' ) as $key ) {
 			// Value `$args[ $key ]` can be a natural number and a numeric string.
 			// We should skip empty string values, but continue working with '0'.
 			if ( ! is_array( $args[ $key ] ) && ( ! is_numeric( $args[ $key ] ) || 0 === $args[ $key ] ) ) {
@@ -351,21 +414,59 @@ class WPForms_Entry_Handler extends WPForms_DB {
 				$ids = (int) $args[ $key ];
 			}
 
-			$where[ 'arg_' . $key ] = "`{$key}` IN ( {$ids} )";
+			$where[ 'arg_' . $key ] = "{$this->table_name}.{$key} IN ( {$ids} )";
 		}
 
 		// Allowed string arg items.
-		$keys = array( 'status', 'type', 'user_uuid' );
-		foreach ( $keys as $key ) {
+		foreach ( array( 'status', 'type', 'user_uuid' ) as $key ) {
 
 			if ( '' !== $args[ $key ] ) {
-				$where[ 'arg_' . $key ] = "`{$key}` = '" . esc_sql( $args[ $key ] ) . "'";
+				$where[ 'arg_' . $key ] = "{$this->table_name}.{$key} = '" . esc_sql( $args[ $key ] ) . "'";
 			}
 		}
 
+		// Processing value and value_compare.
+		if ( ! empty( $args['value'] ) && ! empty( $args['value_compare'] ) ) {
+			switch ( $args['value_compare'] ) {
+				case '': // Preserving backward compatibility.
+				case 'is':
+					$where['arg_value'] = "{$fields_table}.value = '" . esc_sql( $args['value'] ) . "'";
+					break;
+
+				case 'is_not':
+					$where['arg_value'] = "{$fields_table}.value <> '" . esc_sql( $args['value'] ) . "'";
+					break;
+
+				case 'contains':
+					$where['arg_value'] = "{$fields_table}.value LIKE '%" . esc_sql( $args['value'] ) . "%'";
+					break;
+
+				case 'contains_not':
+					$where['arg_value'] = "{$fields_table}.value NOT LIKE '%" . esc_sql( $args['value'] ) . "%'";
+					break;
+			}
+		}
+
+		if ( empty( $args['value'] ) && ! empty( $args['value_compare'] ) ) {
+			// Empty value should be allowed in case certain comparisons are used.
+			switch ( $args['value_compare'] ) {
+				case 'is':
+					$where['arg_value'] = "{$fields_table}.value = ''";
+					break;
+
+				case 'is_not':
+					$where['arg_value'] = "{$fields_table}.value <> ''";
+					break;
+			}
+		}
+
+		if ( isset( $args['field_id'] ) && is_numeric( $args['field_id'] ) ) {
+			$args['field_id']      = (int) $args['field_id'];
+			$where['arg_field_id'] = "{$fields_table}.field_id = '{$args['field_id']}'";
+		}
+
 		// Process dates.
-		$keys = array( 'date', 'date_modified' );
-		foreach ( $keys as $key ) {
+		foreach ( array( 'date', 'date_modified' ) as $key ) {
 			if ( empty( $args[ $key ] ) ) {
 				continue;
 			}
@@ -376,8 +477,8 @@ class WPForms_Entry_Handler extends WPForms_DB {
 				$date_end   = wpforms_get_day_period_date( 'end_of_day', strtotime( $args[ $key ][1] ) );
 
 				if ( ! empty( $date_start ) && ! empty( $date_end ) ) {
-					$where[ 'arg_' . $key . '_start' ] = "`{$key}` >= '{$date_start}'";
-					$where[ 'arg_' . $key . '_end' ]   = "`{$key}` <= '{$date_end}'";
+					$where[ 'arg_' . $key . '_start' ] = "{$this->table_name}.{$key} >= '{$date_start}'";
+					$where[ 'arg_' . $key . '_end' ]   = "{$this->table_name}.{$key} <= '{$date_end}'";
 				}
 			} elseif ( is_string( $args[ $key ] ) ) {
 				/*
@@ -390,8 +491,8 @@ class WPForms_Entry_Handler extends WPForms_DB {
 				$date_end   = wpforms_get_day_period_date( 'end_of_day', $timestamp );
 
 				if ( ! empty( $date_start ) && ! empty( $date_end ) ) {
-					$where[ 'arg_' . $key . '_start' ] = "`{$key}` >= '{$date_start}'";
-					$where[ 'arg_' . $key . '_end' ]   = "`{$key}` <= '{$date_end}'";
+					$where[ 'arg_' . $key . '_start' ] = "{$this->table_name}.{$key} >= '{$date_start}'";
+					$where[ 'arg_' . $key . '_end' ]   = "{$this->table_name}.{$key} <= '{$date_end}'";
 				}
 			}
 		}
@@ -409,6 +510,7 @@ class WPForms_Entry_Handler extends WPForms_DB {
 		 * Modify the ORDER BY.
 		 */
 		$args['orderby'] = ! array_key_exists( $args['orderby'], $this->get_columns() ) ? $this->primary_key : $args['orderby'];
+		$args['orderby'] = "{$this->table_name}.{$args['orderby']}";
 
 		if ( 'ASC' === strtoupper( $args['order'] ) ) {
 			$args['order'] = 'ASC';
@@ -429,30 +531,46 @@ class WPForms_Entry_Handler extends WPForms_DB {
 		 * Retrieve the results.
 		 */
 
-		if ( true === $count ) {
+		$sql_from = $this->table_name;
 
+		// Add a LEFT OUTER JOIN for retrieve a notes count.
+		if ( true === $args['notes_count'] ) {
+			$meta_table = wpforms()->entry_meta->table_name;
+			$sql_from  .= ' LEFT JOIN';
+			$sql_from  .= " ( SELECT {$meta_table}.entry_id AS meta_entry_id, COUNT({$meta_table}.id) AS notes_count";
+			$sql_from  .= " FROM {$meta_table}";
+			$sql_from  .= " WHERE {$meta_table}.type = 'note'";
+			$sql_from  .= ' GROUP BY meta_entry_id )';
+			$sql_from  .= " notes_counts ON notes_counts.meta_entry_id = {$this->table_name}.entry_id";
+
+			// Changed the ORDER BY - notes count sorting support.
+			if ( "{$this->table_name}.notes_count" === $args['orderby'] ) {
+				$args['orderby'] = 'notes_counts.notes_count';
+			}
+		}
+
+		if ( ! empty( $args['value'] ) || ! empty( $args['value_compare'] ) ) {
+			$sql_from .= " JOIN {$fields_table} ON {$this->table_name}.entry_id={$fields_table}.entry_id";
+		}
+
+		if ( true === $count ) {
 			// @codingStandardsIgnoreStart
-			$results = absint( $wpdb->get_var(
-				"SELECT COUNT({$this->primary_key})
-				FROM {$this->table_name}
+			return absint( $wpdb->get_var(
+				"SELECT COUNT({$this->table_name}.{$this->primary_key}) 
+				FROM {$sql_from}
 				WHERE {$where_sql};"
 			) );
 			// @codingStandardsIgnoreEnd
-
-		} else {
-
-			// @codingStandardsIgnoreStart
-			$results = $wpdb->get_results(
-				"SELECT {$select}
-				FROM {$this->table_name}
-				WHERE {$where_sql}
-				ORDER BY {$args['orderby']} {$args['order']}
-				LIMIT {$args['offset']}, {$args['number']};"
-			);
-			// @codingStandardsIgnoreEnd
 		}
 
-		return $results;
+		$sql = "SELECT {$select} 
+			FROM {$sql_from}";
+
+		$sql .= " WHERE {$where_sql} 
+			ORDER BY {$args['orderby']} {$args['order']} 
+			LIMIT {$args['offset']}, {$args['number']};";
+
+		return $wpdb->get_results( $sql ); // phpcs:ignore
 	}
 
 	/**
